requirements.txt
streamlit>=1.32,<1.38
yfinance>=0.2.40
pandas>=2.2.2
numpy==1.26.4
scikit-learn>=1.4.2,<1.6
tensorflow-cpu==2.16.1
lxml>=5.2.1
html5lib>=1.1
requests>=2.32.0
certifi>=2024.2.2
tqdm>=4.66.4
